refactor: Extract disease tools (OMIM) into separate module
- Create src/tools/disease_tools.py with 3 OMIM functions
- Move disease/OMIM logic from server.py:
  - get_omim_by_mim_number()
  - get_omim_by_gene_symbol()
  - get_omim_variant()
- Add register_tools() function for clean registration
- Update server.py to import and register disease tools

Benefits:
- Groups OMIM-related queries in one place
- Reduced server.py by ~80 lines
- Easy to add other disease databases (HPO, Orphanet, etc.)
- Better code organization

All existing functionality preserved - backward compatible.
