+++ server.py
@@ -16,7 +16,7 @@ from mcp.server.fastmcp import FastMCP
 from src.utils.api_client import fetch_marrvel_data
 
 # Import tool modules
-from src.tools import gene_tools, variant_tools
+from src.tools import gene_tools, variant_tools, disease_tools
 
 # Initialize FastMCP server
 mcp = FastMCP("MARRVEL")
@@ -24,6 +24,7 @@ mcp = FastMCP("MARRVEL")
 # Register tools
 gene_tools.register_tools(mcp)
 variant_tools.register_tools(mcp)
+disease_tools.register_tools(mcp)
 
 
 # ============================================================================
@@ -48,92 +49,13 @@ variant_tools.register_tools(mcp)
 
 
 # ============================================================================
-# DISEASE TOOLS (OMIM)
 # ============================================================================
-
-@mcp.tool()
-async def get_omim_by_mim_number(mim_number: str) -> str:
-    """
-    Retrieve OMIM (Online Mendelian Inheritance in Man) entry by MIM number.
-    
-    OMIM is a comprehensive database of human genes and genetic disorders.
-    
-    Args:
-        mim_number: OMIM MIM number (e.g., "191170" for Treacher Collins syndrome)
-        
-    Returns:
-        JSON string with OMIM entry:
-        - Disease/phenotype description
-        - Clinical features
-        - Inheritance pattern
-        - Molecular genetics
-        - Allelic variants
-        
-    Example:
-        get_omim_by_mim_number("191170")  # Treacher Collins syndrome
-        get_omim_by_mim_number("114480")  # Breast cancer (BRCA1)
-    """
-    try:
-        data = await fetch_marrvel_data(f"/omim/mimNumber/{mim_number}")
-        return str(data)
-    except httpx.HTTPError as e:
-        return f"Error fetching OMIM data: {str(e)}"
-
-
-@mcp.tool()
-async def get_omim_by_gene_symbol(gene_symbol: str) -> str:
-    """
-    Find all OMIM diseases associated with a gene symbol.
-    
-    This tool retrieves all OMIM entries (diseases, phenotypes) that are
-    associated with a particular gene.
-    
-    Args:
-        gene_symbol: Official gene symbol (e.g., "TP53", "BRCA1", "CFTR")
-        
-    Returns:
-        JSON string with list of OMIM diseases including:
-        - MIM numbers
-        - Disease names
-        - Inheritance patterns
-        - Gene-disease relationships
-        
-    Example:
-        get_omim_by_gene_symbol("TP53")  # Li-Fraumeni syndrome
-        get_omim_by_gene_symbol("BRCA1")  # Breast/ovarian cancer
-        get_omim_by_gene_symbol("CFTR")  # Cystic fibrosis
-    """
-    try:
-        data = await fetch_marrvel_data(f"/omim/gene/symbol/{gene_symbol}")
-        return str(data)
-    except httpx.HTTPError as e:
-        return f"Error fetching OMIM data: {str(e)}"
-
-
-@mcp.tool()
-async def get_omim_variant(gene_symbol: str, variant: str) -> str:
-    """
-    Query OMIM for specific variant information.
-    
-    Get OMIM data for a specific variant in a gene, including disease
-    associations and clinical significance.
-    
-    Args:
-        gene_symbol: Gene symbol (e.g., "TP53")
-        variant: Variant description (e.g., "p.R248Q", "c.743G>A")
-        
-    Returns:
-        JSON string with variant-specific OMIM information
-        
-    Example:
-        get_omim_variant("TP53", "p.R248Q")
-        get_omim_variant("BRCA1", "p.C61G")
-    """
-    try:
-        data = await fetch_marrvel_data(f"/omim/gene/symbol/{gene_symbol}/variant/{variant}")
-        return str(data)
-    except httpx.HTTPError as e:
-        return f"Error fetching OMIM data: {str(e)}"
+# DISEASE TOOLS (OMIM) - Now imported from src.tools.disease_tools
+# ============================================================================
+# The following disease tools are now registered from the disease_tools module:
+# - get_omim_by_mim_number
+# - get_omim_by_gene_symbol
+# - get_omim_variant
 
 
 # ============================================================================
